# web/src/data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "3D Asset Server API",
    "version": "0.1.0",
    "summary": "One search API for free and paid 3D models, PBR materials, textures, HDRIs and game assets.",
    "description": "Search 17 asset sites at once (Poly Haven, ambientCG, Kenney, BlenderKit, CGTrader, itch.io and more),\nget licences and file lists, and download glTF/FBX/Blend models, PBR texture maps and HDRIs.\n\n**Typical flow:** `GET /v1/search` → `GET /v1/assets/{id}` → `GET /v1/assets/{id}/download`.\n\n**Authentication:** the public server at 3d.shep.bot is open. Self-hosted servers can require a key\n(`ASSET_SERVER_API_KEY`); send it as `Authorization: Bearer <key>` or `x-api-key: <key>`.\n\n**For AI agents:** the same capabilities are exposed as MCP tools at `/mcp` (Streamable HTTP).\nSee [Use with coding agents](/docs/mcp).\n\n**Rate limits:** 120 requests per minute per client on `/v1/*` and `/mcp`. Every response carries\n`RateLimit-Policy` and `RateLimit` (IETF RateLimit header fields) plus `RateLimit-Limit`,\n`RateLimit-Remaining` and `RateLimit-Reset`; a `429` adds `Retry-After` (seconds).\n\n**Versioning:** the major version is in the path (`/v1`). Breaking changes only ship in a new major\nversion; deprecated endpoints announce themselves with `Deprecation` and `Sunset` headers at least\n90 days ahead. See [Versioning & deprecation](/docs/api/versioning).",
    "contact": {
      "name": "3D Asset Server",
...
    },
    "license": {
...
    }
  },
...
  "servers": [
...
  "security": [
    {},
    {
      "bearer": []
    },
    {
      "apiKey": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Only when the server sets ASSET_SERVER_API_KEY."
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Same key, as a header."
      }
    },
    "headers": {
      "RateLimit": {
        "description": "Current quota, e.g. `\"default\";r=117;t=42` (r = requests remaining, t = seconds until reset).",
        "schema": {
          "type": "string"
        }
      },
      "RateLimit-Policy": {
        "description": "Quota policy, e.g. `\"default\";q=120;w=60` (q = requests per window, w = window seconds).",
        "schema": {
          "type": "string"
        }
      },
      "RateLimit-Limit": {
        "description": "Requests allowed per window.",
        "schema": {
          "type": "integer"
        }
      },
...
    },
...
  },
  "paths": {
    "/v1/search": {
      "get": {
        "operationId": "searchAssets",
        "tags": [
          "Search"
        ],
        "summary": "Search every source",
        "description": "Fans out to every enabled source in parallel (each with its own timeout), then merges, ranks and de-duplicates. A slow or failing source never fails the search: it shows up in `providers` with its status.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Short, concrete query.",
            "schema": {
              "type": "string"
            },
            "example": "low poly tree"
          },
          {
            "name": "type",
            "in": "query",
            "description": "Comma-separated asset types: model, texture, material, hdri, sprite, ui, audio, font, pack, other.",
            "schema": {
              "type": "string"
            },
            "example": "model"
          },
          {
            "name": "providers",
            "in": "query",
            "description": "Comma-separated source ids (see /v1/providers).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "free",
            "in": "query",
            "description": "Only free assets.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "downloadable",
            "in": "query",
            "description": "Only assets this server can download directly.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max results (default 24).",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Per-source offset for paging (e.g. 24 for page 2).",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked results plus a per-source report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "headers": {
              "RateLimit": {
                "$ref": "#/components/headers/RateLimit"
              },
              "RateLimit-Policy": {
                "$ref": "#/components/headers/RateLimit-Policy"
              },
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            }
          },
          "400": {
            "description": "Invalid parameters or unknown source id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "RateLimit": {
                "$ref": "#/components/headers/RateLimit"
              },
              "RateLimit-Policy": {
                "$ref": "#/components/headers/RateLimit-Policy"
              },
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            }
          },
          "401": {
            "description": "API key required (self-hosted servers only).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "RateLimit": {
                "$ref": "#/components/headers/RateLimit"
              },
              "RateLimit-Policy": {
                "$ref": "#/components/headers/RateLimit-Policy"
              },
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/assets/{id}": {
      "get": {
        "operationId": "getAsset",
        "tags": [
          "Assets"
        ],
        "summary": "Asset details with files",
        "description": "Full metadata, licence and every downloadable file for one asset.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Asset id `<provider>:<nativeId>`, URL-encoded (`:` may stay as is). Take it from a search result.",
            "schema": {
              "type": "string"
            },
            "example": "polyhaven:ArmChair_01"
          }
        ],
        "responses": {
          "200": {
            "description": "The asset and its files.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDetails"
                }
              }
            },
            "headers": {
              "RateLimit": {
                "$ref": "#/components/headers/RateLimit"
              },
              "RateLimit-Policy": {
                "$ref": "#/components/headers/RateLimit-Policy"
              },
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            }
          },
          "404": {
            "description": "Unknown asset.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "RateLimit": {
                "$ref": "#/components/headers/RateLimit"
              },
              "RateLimit-Policy": {
                "$ref": "#/components/headers/RateLimit-Policy"
              },
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "502": {
            "description": "The source site failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "RateLimit": {
                "$ref": "#/components/headers/RateLimit"
              },
              "RateLimit-Policy": {
                "$ref": "#/components/headers/RateLimit-Policy"
              },
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            }
          }
        }
      }
    },
    "/v1/assets/{id}/files": {
      "get": {
        "operationId": "selectFiles",
        "tags": [
          "Assets"
        ],
        "summary": "Pick the right files",
        "description": "Smart selection for a format and resolution (e.g. glTF with its .bin and textures, or a 2k PBR map set).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Asset id `<provider>:<nativeId>`, URL-encoded (`:` may stay as is). Take it from a search result.",
            "schema": {
              "type": "string"
            },
            "example": "polyhaven:ArmChair_01"
          },
          {
            "name": "format",
            "in": "query",
            "description": "Preferred format or package. Closest match wins; omitted = best default for the asset type.",
            "schema": {
              "type": "string",
              "examples": [
                "glb",
                "gltf",
                "fbx",
                "blend",
                "obj",
                "usd",
                "hdr",
                "exr",
                "jpg",
                "png",
                "zip"
              ]
            }
          },
          {
            "name": "resolution",
            "in": "query",
            "description": "Texture/HDRI resolution. The closest available is used (default 2k).",
            "schema": {
              "type": "string",
              "examples": [
                "1k",
                "2k",
                "4k",
                "8k"
              ]
            }
          },
          {
            "name": "maps",
            "in": "query",
            "description": "Comma-separated texture map types for map sets, e.g. `diff,nor_gl,rough,ao`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "all",
            "in": "query",
            "description": "Return every file instead of the smart selection.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Selected files and total size.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileSelection"
                }
              }
            },
            "headers": {
              "RateLimit": {
                "$ref": "#/components/headers/RateLimit"
              },
              "RateLimit-Policy": {
                "$ref": "#/components/headers/RateLimit-Policy"
              },
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            }
          },
          "404": {
            "description": "Unknown asset.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "RateLimit": {
                "$ref": "#/components/headers/RateLimit"
              },
              "RateLimit-Policy": {
                "$ref": "#/components/headers/RateLimit-Policy"
              },
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
    "/openapi.json": {
      "get": {
        "operationId": "openapi",
        "tags": [
          "System"
        ],
        "summary": "This OpenAPI document",
        "security": [
          {}
        ],
        "responses": {
          "200": {
            "description": "OpenAPI 3.1",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  }
}